refactor: 변수명, 함수명 리팩터링
