+++ src/render/renderWinningLottoForm.js
