+++ src/style/index.css
@@ -4,7 +4,7 @@
 @import url(./footer.css);
 @import url(./purchaseAmountForm.css);
 @import url(./purchaseLotto.css);
-@import url(./winningForm.css);
+@import url(./winningLottoForm.css);
 @import url(./statistics.css);
 
 #app {
